fix: pr feedbacks and temporary workaround for #1081
